+++ lib/AppConfig.php
@@ -320,6 +320,13 @@ class AppConfig {
     public $_customizationPlugins = "customization_plugins";
 
     /**
+     * The config key for the interval of editors availability check by cron
+     *
+     * @var string
+     */
+    private $_editors_check_interval = "editors_check_interval";
+
+    /**
      * @param string $AppName - application name
      */
     public function __construct($AppName) {
@@ -1144,18 +1151,39 @@ class AppConfig {
     /**
      * Get the jwt header setting
      *
+     * @param bool $origin - take origin
+     *
      * @return string
      */
-    public function JwtHeader() {
-        if ($this->UseDemo()) {
+    public function JwtHeader($origin = false) {
+        if (!$origin && $this->UseDemo()) {
             return $this->DEMO_PARAM["HEADER"];
         }
 
-        $header = $this->GetSystemValue($this->_jwtHeader);
+        $header = $this->config->getAppValue($this->appName, $this->_jwtHeader, "");
         if (empty($header)) {
+            $header = $this->GetSystemValue($this->_jwtHeader);
+        }
+        if (!$origin && empty($header)) {
             $header = "Authorization";
         }
         return $header;
+    }
+
+    /**
+     * Save the jwtHeader setting
+     *
+     * @param string $value - jwtHeader
+     */
+    public function SetJwtHeader($value) {
+        $value = trim($value);
+        if (empty($value)) {
+            $this->logger->info("Clear header key", ["app" => $this->appName]);
+        } else {
+            $this->logger->info("Set header key " . $value, ["app" => $this->appName]);
+        }
+
+        $this->config->setAppValue($this->appName, $this->_jwtHeader, $value);
     }
 
     /**
@@ -1255,6 +1283,20 @@ class AppConfig {
     }
 
     /**
+     * Get the editors check interval
+     *
+     * @return int
+     */
+    public function GetEditorsCheckInterval() {
+        $interval = $this->GetSystemValue($this->_editors_check_interval);
+
+        if (empty($interval)) {
+            $interval = 60*60*24;
+        }
+        return (integer)$interval;
+    }
+
+    /**
      * Additional data about formats
      *
      * @var array
@@ -1306,4 +1348,15 @@ class AppConfig {
         "SECRET" => "sn2puSUF7muF5Jas",
         "TRIAL" => 30
     ];
+
+    private $linkToDocs = "https://www.onlyoffice.com/docs-registration.aspx?referer=nextcloud";
+
+    /**
+     * Get link to Docs Cloud
+     *
+     * @return string
+     */
+    public function GetLinkToDocs() {
+        return $this->linkToDocs;
+    }
 }