+++ lib/Controller/SettingsController.php
@@ -110,6 +110,7 @@ class SettingsController extends Controller {
             "storageUrl" => $this->config->GetStorageUrl(),
             "verifyPeerOff" => $this->config->GetVerifyPeerOff(),
             "secret" => $this->config->GetDocumentServerSecret(true),
+            "jwtHeader" => $this->config->JwtHeader(true),
             "demo" => $this->config->GetDemoData(),
             "currentServer" => $this->urlGenerator->getAbsoluteURL("/"),
             "formats" => $this->config->FormatsSetting(),
@@ -132,7 +133,8 @@ class SettingsController extends Controller {
             "tagsEnabled" => \OC::$server->getAppManager()->isEnabledForUser("systemtags"),
             "reviewDisplay" => $this->config->GetCustomizationReviewDisplay(),
             "theme" => $this->config->GetCustomizationTheme(),
-            "templates" => $this->GetGlobalTemplates()
+            "templates" => $this->GetGlobalTemplates(),
+            "linkToDocs" => $this->config->GetLinkToDocs()
         ];
         return new TemplateResponse($this->appName, "settings", $data, "blank");
     }
@@ -145,6 +147,7 @@ class SettingsController extends Controller {
      * @param string $storageUrl - Nextcloud address available from document server
      * @param bool $verifyPeerOff - parameter verification setting
      * @param string $secret - secret key for signature
+     * @param string $jwtHeader - jwt header
      * @param bool $demo - use demo server
      *
      * @return array
@@ -154,6 +157,7 @@ class SettingsController extends Controller {
                                     $storageUrl,
                                     $verifyPeerOff,
                                     $secret,
+                                    $jwtHeader,
                                     $demo
                                     ) {
         $error = null;
@@ -165,6 +169,7 @@ class SettingsController extends Controller {
             $this->config->SetVerifyPeerOff($verifyPeerOff);
             $this->config->SetDocumentServerInternalUrl($documentserverInternal);
             $this->config->SetDocumentServerSecret($secret);
+            $this->config->SetJwtHeader($jwtHeader);
         }
         $this->config->SetStorageUrl($storageUrl);
 
@@ -184,6 +189,7 @@ class SettingsController extends Controller {
             "documentserverInternal" => $this->config->GetDocumentServerInternalUrl(true),
             "storageUrl" => $this->config->GetStorageUrl(),
             "secret" => $this->config->GetDocumentServerSecret(true),
+            "jwtHeader" => $this->config->JwtHeader(true),
             "error" => $error,
             "version" => $version,
             ];