+++ templates/settings.php
@@ -62,6 +62,9 @@
             </a>
         </p>
         <div id="onlyofficeSecretPanel" class="onlyoffice-hide">
+            <p class="onlyoffice-header"><?php p($l->t("Authorization header (leave blank to use default header)")) ?></p>
+            <p><input id="onlyofficeJwtHeader" value="<?php p($_["jwtHeader"]) ?>" placeholder="Authorization" type="text"></p>
+
             <p class="onlyoffice-header"><?php p($l->t("ONLYOFFICE Docs address for internal requests from the server")) ?></p>
             <p><input id="onlyofficeInternalUrl" value="<?php p($_["documentserverInternal"]) ?>" placeholder="https://<documentserver>/" type="text"></p>
 
@@ -86,6 +89,25 @@
             <?php } else { ?>
             <em><?php p($l->t("The 30-day test period is over, you can no longer connect to demo ONLYOFFICE Docs server.")) ?></em>
             <?php } ?>
+        </div>
+    </div>
+
+    <div id="onlyofficeDocsCloudBannerWrapper">
+        <div id="onlyofficeDocsCloudBannerPicWrapper">
+            <div id="onlyofficeDocsCloudBannerPic">
+
+            </div>
+        </div>
+        <div id="onlyofficeDocsCloudBannerContent">
+            <div id="onlyofficeDocsCloudBannerContentText">
+                <p><h2>ONLYOFFICE Docs Cloud</h2></p>
+                <p><?php p($l->t("Easily launch the editors in the cloud without downloading and installation")) ?></p>
+            </div>
+            <div id="onlyofficeDocsCloudBannerContentButtonWrapper">
+                <a id="onlyofficeDocsCloudBannerContentButton" class="button" href="<?php p($_["linkToDocs"]) ?>" target="_blank">
+                    <?php p($l->t("Get Now")) ?>
+                </a>
+            </div>
         </div>
     </div>